refactor(front): rename Job{Steps→Progress}.vue
To avoid confusion with the notion of jobs steps existing in Slurm.
