+++ frontend/src/components/jobs/JobProgress.vue
